requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
supabase==2.3.1
python-dotenv==1.0.0
openai==1.11.0
pydantic==2.5.3
python-multipart==0.0.6